Add files via upload
Use Naive Bayes on a Text Classification Task
Using the following training set of pairs (sentence, class): 
("Your limitation—it’s only your imagination", Inspiring),
("Push yourself, because no one else is going to do it for you", Inspiring),
("Sometimes later becomes never, do it now", Inspiring)
("Great things never come from comfort zones", Inspiring)
("It's great to be yourself", Dull)
("A goal is a dream with a plan", Dull)

Classify the sentence: "Dream it. Wish it. Do it." as either Inspiring or Dull using the Naive Bayes method (as in Example 13.1).